refactor: User interface and imports
#8
